[#1087] Split logic of ServicePanel into separate components.
