+++ frontend/src/service-panel/Button.tsx
@@ -1,0 +1,18 @@
+import * as React from 'react';
+
+import './Button.less';
+
+interface ButtonProps {
+  onClick(e: React.MouseEvent): void;
+}
+
+const Button: React.StatelessComponent<ButtonProps> = ({ onClick }) => (
+  <button
+    className='ServicePanelButton'
+    onClick={onClick}
+  >
+    <span className='icon icon-cog' />
+  </button>
+);
+
+export default Button;
