+++ frontend/src/service-panel/VpTitle.tsx
@@ -1,0 +1,9 @@
+import * as React from 'react';
+
+const VpTitle: React.StatelessComponent<{}> = () => (
+  <h1 className='vp-header'>
+    VersionPress
+  </h1>
+);
+
+export default VpTitle;
